Updated standard structure

+++ scan.cpp
@@ -1,4 +1,5 @@
 #include "stdafx.h"
+#include "scan.h"
 
 #ifdef _MSC_VER
 #if _MSC_VER==1200
@@ -14,9 +15,6 @@
 
 int scanf_result = 0;
 
-/// <summary>
-/// 扫描 0001-9999 的数字。
-/// </summary>
 void ScanShort(const char* message, short *i)
 {
 	printf(message);
@@ -33,9 +31,6 @@ void ScanShort(const char* message, short *i)
 	}
 }
 
-/// <summary>
-/// 扫描32位整形。
-/// </summary>
 void ScanInt(const char* message, int *i)
 {
 	printf(message);
@@ -53,9 +48,6 @@ void ScanInt(const char* message, int *i)
 	}
 }
 
-/// <summary>
-/// 扫描双精度浮点。
-/// </summary>
 void ScanDouble(const char* message, double *i)
 {
 	printf(message);
@@ -73,10 +65,7 @@ void ScanDouble(const char* message, double *i)
 	}
 }
 
-/// <summary>
-/// 扫描len为长度的字符串进入buffer。
-/// </summary>
-void ScanText(const char* message, char *buffer, int len)
+void ScanText(const char* message, char *buffer, size_t len)
 {
 	memset(buffer, 0x00, sizeof(char) * len);
 	printf(message);
@@ -90,9 +79,6 @@ void ScanText(const char* message, char *buffer, int len)
 		ScanText(message, buffer, len);
 }
 
-/// <summary>
-/// 扫描a-e或者1-2连续ASCII的选项
-/// </summary>
 void ScanOption(const char* message, const char min, const char max, char *c)
 {
 	printf(message);
@@ -112,9 +98,6 @@ void ScanOption(const char* message, const char min, const char max, char *c)
 #define isTrue(c) ((c=='y')||(c=='Y'))
 #define isFalse(c) ((c=='n')||(c=='N'))
 
-/// <summary>
-/// 扫描是 / 否选项。
-/// </summary>
 bool ScanBoolean(const char* message)
 {
 	printf(message);
